refactor: expose deploy function
